# backend/openapi.yaml
openapi: 3.0.3
info:
  title: MayaTrail API
  version: 1.0.0
  description: |
    Backend API for the MayaTrail cloud-security simulation platform.

    **Authentication**: All endpoints (except `/api/auth/register` and `/api/auth/login`)
    require a JWT Bearer token obtained via the login endpoint.
  contact:
    name: MayaTrail Team
servers:
  - url: http://localhost:8000
    description: Local development (Docker Compose)

# ---------------------------------------------------------------------------
# Security
# ---------------------------------------------------------------------------
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Pass the access token returned by POST /api/auth/login
        in the `Authorization: Bearer <token>` header.

  # -----------------------------------------------------------------------
  # Reusable schemas
  # -----------------------------------------------------------------------
  schemas:
...
    SimulationRun:
      type: object
      properties:
        id:
          type: string
          format: uuid
        stack:
          type: string
          format: uuid
          description: UUID of the Stack used for this run.
        module:
          type: string
          enum:
            - attach_role_policy
            - enumeration
            - eventual_consistency
            - s3_initial_access
            - s3_kms_encryption
          example: enumeration
        status:
          type: string
          enum: [pending, running, completed, failed]
          example: pending
        stdout:
          type: string
          example: ""
        stderr:
          type: string
          example: ""
        triggered_by:
          type: string
          description: Username of the user who triggered the run.
          example: himan10
        started_at:
          type: string
          format: date-time
          nullable: true
        completed_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
    SimulationModule:
      type: object
      properties:
        id:
          type: integer
          description: Numeric ID used to trigger this simulation.
          example: 2
        name:
          type: string
          description: Internal module name.
          example: enumeration
        description:
          type: string
          description: Human-readable description of the simulation.
          example: IAM policy simulator / service enumeration
    TriggerSimulationRequest:
      type: object
      required: [stack_id, module_id]
      properties:
        stack_id:
          type: string
          format: uuid
          description: UUID of the target Stack (must be READY).
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
        module_id:
          type: integer
          enum: [1, 2, 3, 4, 5]
          description: Numeric ID of the simulation module (see GET /api/simulations/modules/).
          example: 2
    TriggerSimulationResponse:
      type: object
      properties:
        run:
          $ref: "#/components/schemas/SimulationRun"
        task_id:
          type: string
          example: d4e5f6a7-b8c9-0123-4567-890abcdef123

    # ---- Logs -----------------------------------------------------------
...
    ValidationError:
      type: object
      additionalProperties:
        type: array
        items:
          type: string
      example:
        username: ["This field is required."]
    DetailError:
      type: object
      properties:
        detail:
          type: string
          example: "Authentication credentials were not provided."

  # -----------------------------------------------------------------------
  # Reusable responses
  # -----------------------------------------------------------------------
  responses:
    Unauthorized:
      description: Authentication credentials were not provided or are invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/DetailError"

# ---------------------------------------------------------------------------
# Paths
# ---------------------------------------------------------------------------
paths:
...
  /api/simulations/modules/:
    get:
      tags: [Simulations]
      summary: List all available simulation modules
      operationId: simulationsModules
      description: >
        Returns the catalogue of simulation modules with their numeric IDs,
        internal names, and human-readable descriptions. Use the `id` from
        this response as the `module_id` when triggering a simulation run.
      responses:
        "200":
          description: Array of available simulation modules.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SimulationModule"
              example:
                - id: 1
                  name: attach_role_policy
                  description: Privilege escalation via AttachRolePolicy
                - id: 2
                  name: enumeration
                  description: IAM policy simulator / service enumeration
                - id: 3
                  name: eventual_consistency
                  description: Eventual consistency attack
                - id: 4
                  name: s3_initial_access
                  description: S3 basic access & data exfiltration
                - id: 5
                  name: s3_kms_encryption
                  description: S3 KMS ransomware simulation
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/simulations/run/:
    post:
      tags: [Simulations]
      summary: Trigger a new simulation run
      operationId: simulationsRun
      description: >
        Creates a SimulationRun record and enqueues a Celery task.
        The referenced stack must belong to the user and be in READY status.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TriggerSimulationRequest"
      responses:
        "201":
          description: Run created and task enqueued.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TriggerSimulationResponse"
        "400":
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
        "404":
          description: Stack not found or doesn't belong to user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DetailError"
        "409":
          description: Stack is not in READY state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DetailError"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ======================= Logs ==========================================
...